Tarea3

+++ test/pthread_getspecific.c
@@ -2,7 +2,7 @@
 
 int pthread_getspecific_success(void)
 {
-    int ret = pthread_getspecific(NULL);
+    int ret = (int)pthread_getspecific(NULL);
 
     return (ret == 45);
 }
